requirements.txt
# Основные зависимости Django
Django>=4.2,<5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# WSGI сервер для production
gunicorn>=21.2.0

# tinymce
django-tinymce==4.1.0

# Переводы и интернационализация
django-parler>=2.3

# Теги
django-taggit>=4.0.0

# SEO метаданные
django-meta>=2.4.0

# Изображения и миниатюры
Pillow>=10.0.0
easy-thumbnails>=2.8.0
django-filer>=3.1.0
django-mptt>=0.15.0

# Статические файлы
# whitenoise>=6.5.0

# Кеширование (опционально)
django-redis>=5.3.0

# Dev зависимости (опционально)
django-browser-reload>=1.8.0

# Дополнительные зависимости для работы системы
requests>=2.31.0  # ДОБАВЛЕНО: для HTTP запросов (используется в core.views)

# Удалено: django-ckeditor - больше не нужен
# Удалено: django-ckeditor-uploader - больше не нужен
pytz>=2023.3